Fixes sorting on bin changes
