+++ src/utils/binning-helpers.js
@@ -124,3 +124,16 @@ export const createBinParams = (chart, binParams) => {
   chart.group().binParams(parsedBinParams)
   return chart
 }
+
+export const getFirstNonNullDatumForAxis = (data, axisType) => {
+  const keyName = `key${axisType === "x" ? "0" : "1"}`
+  return (
+    data &&
+    Array.isArray(data) &&
+    data.find((datum = {}) => {
+      const keyVal = datum[keyName]
+      const value = Array.isArray(keyVal) ? keyVal[0] : keyVal
+      return value !== null
+    })
+  )
+}
